+++ Cargo.toml
@@ -3,8 +3,9 @@ resolver = "2"
 members = [
   "node",
   "node/data-extension/worker",
+  "node/data-extension/wasm/executor",
   "node/data-extension/wasm/utils",
-  "node/data-extension/wasm/gen/bt",
+  "node/data-extension/wasm/test-wasm",
   "libs/mocks",
   "libs/primitives",
   "libs/proofs",
@@ -90,7 +91,7 @@ impl-trait-for-tuples = "0.2.1"
 num-traits = { version = "0.2", default-features = false }
 async-channel = "1.8.0"
 futures-channel = "0.3.29"
-wasmtime = "17.0.1"
+wasmtime = { version = "17.0.1", default-features = false }
 
 # Cumulus
 cumulus-pallet-aura-ext = { git = "https://github.com/paritytech/cumulus", default-features = false, branch = "polkadot-v0.9.43" }
@@ -318,6 +319,7 @@ vergen = "3.0.4"
 # Data extension
 data-extension-worker = { path = "node/data-extension/worker", default-features = false }
 wasm-utils = { path = "node/data-extension/wasm/utils", default-features = false }
+wasm-executor = { path = "node/data-extension/wasm/executor", default-features = false }
 
 #
 # Cargo patch rules for all the paritytech/ based crates.
