+++ node/data-extension/wasm/executor/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "wasm-executor"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+documentation.workspace = true
+
+[dependencies]
+thiserror = { workspace = true, default-features = true }
+wasmtime = { workspace = true, default-features = true }
+wasm-utils = { workspace = true, default-features = true }
+
+[dev-dependencies]
+serde = { workspace = true, default-features = true }
+serde_json = { workspace = true, default-features = true }
