+++ node/data-extension/wasm/test-wasm/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wasm-gen-bt"
+name = "test-wasm"
 version = "0.1.0"
 authors.workspace = true
 edition.workspace = true
