--- node/data-extension/worker/src/service/document/mod.rs
@@ -1,1 +1,0 @@
-//TODO(cdamian): CRUD Service for documents
